Implement long message ids

+++ app/src/main/java/de/kuschku/quasseldroid/service/QuasselService.kt
@@ -179,8 +179,8 @@ class QuasselService : DaggerLifecycleService(),
       }
     }
 
-    val clearMessageId = intent.getIntExtra("mark_read_message", -1)
-    if (bufferId != -1 && clearMessageId != -1) {
+    val clearMessageId = intent.getLongExtra("mark_read_message", -1)
+    if (bufferId != -1 && clearMessageId != -1L) {
       sessionManager.session.value?.bufferSyncer?.requestSetLastSeenMsg(bufferId, clearMessageId)
       sessionManager.session.value?.bufferSyncer?.requestMarkBufferAsRead(bufferId)
     }
